# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Instant Loan Pre-Decision API",
    "version": "1.0.0",
    "description": "Accepts a loan application and returns an automated **pre-decision** with stable reason codes.\n\nThis is not the final credit decision. Where an application is referred, a person at the lender\ndecides on criteria this service does not model, and that outcome is a separate record.\n\nThe reasoning behind every field and error code is in `docs/05-api.md`; this document is the\nmachine-readable half, generated from the route schemas so it cannot drift from the code."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A bearer token in the form the service was configured with.\n\nThree separate lists exist — submission, review and audit — and a token valid for one scope\nreturns 403 on another. Use the **Authorize** button; the demo token is in the submission email."
      }
    },
    "schemas": {}
  },
  "paths": {
    "/health/live": {
      "get": {
        "summary": "Liveness",
        "tags": [
          "health"
        ],
        "description": "Touches nothing. A failing liveness probe means \"restart me\", so a probe that checked a dependency would turn a database outage into a container restart loop.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Prometheus metrics",
        "tags": [
          "operations"
        ],
        "description": "Behind the auditor scope, not public: this instance has one public URL and no second bind, so an open endpoint would make the outcome mix and pull volumes world-readable. Nothing scrapes this in v1.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Prometheus text exposition format",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "Prometheus text exposition format"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Readiness",
        "tags": [
          "health"
        ],
        "description": "Verifies the database answers. Returns 503 when it does not, so a load balancer stops routing here without the container being killed.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ready"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/applications": {
      "post": {
        "summary": "Submit an application and receive a pre-decision",
        "tags": [
          "applications"
        ],
        "description": "Synchronous: validate, screen, look the applicant up at the bureau, decide, persist, answer.\n\n`Idempotency-Key` is optional but strongly recommended. Without one, a retry after a network\ntimeout creates a second application — it will NOT create a second bureau enquiry, which is\nprevented separately, but it will create a second record. With one, an identical retry returns\nthe stored response byte for byte and carries `Idempotency-Replayed: true`.\n\n`preDecision.reasonCodes` may be an empty array. That is correct and happens for exactly one\ncase: an approval on the terms applied for, which is not adverse action and owes the applicant\nno explanation. Read `length === 0` rather than treating the field as absent.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitApplicationInput"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "in": "header",
            "name": "idempotency-key",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Submission"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/applications/{id}": {
      "get": {
        "summary": "Read an application and its composed outcome",
        "tags": [
          "applications"
        ],
        "description": "The reason a synchronous API still needs a status call: `MANUAL_REVIEW` is not terminal.\n\n`preDecision` never changes — it is what the engine concluded, and a human disagreeing with it\ndoes not make it untrue. `outcome` is the composed answer and is the only field that moves.\n\nReads are owner-scoped. An unknown id and another client's id return the same 404 body:\ndistinguishing them would turn this endpoint into an oracle confirming which ids are real.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationStatus"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reviews/{applicationId}/close": {
      "post": {
        "summary": "Record the outcome of a manual review",
        "tags": [
          "reviews"
        ],
        "description": "Records what a PERSON concluded, on criteria this service does not model. It is not a review\nworkflow: no queue, no assignment, no SLA, no UI.\n\n`reviewerId` comes from the bearer token and never from the body — a human verdict with no\nattributable actor cannot answer \"could anyone have altered a verdict after the fact?\", and\nletting the caller name themselves is the same hole with extra steps.\n\nThe write is conditional on the review still being PENDING, so two concurrent closes produce\none write and one 409. A closed review is never reopened; a mistaken outcome is a new application.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseReviewInput"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "applicationId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationStatus"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audit/applications/{id}/events": {
      "get": {
        "summary": "The append-only event trail for one application",
        "tags": [
          "audit"
        ],
        "description": "Index 2 is named BUREAU_REPORT_ATTACHED rather than STORED because on the reuse path nothing is written. An audit trail that records a write which did not happen is one nobody can rely on.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEvents"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audit/applications/{id}/chain": {
      "get": {
        "summary": "Verify the hash chain",
        "tags": [
          "audit"
        ],
        "description": "Recomputes every hash and compares. Honest limits, because a verifier that oversells itself is\nworse than none: this detects an edit that got past the append-only trigger. It does NOT detect\na consistent rewrite by someone with full database access, and it does not detect TRUNCATION —\ndelete the last k events and what remains verifies perfectly. `events` is returned so an\nexternal anchor publishing the head hash AND the count would have something to compare against.\nThat anchor is not built.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainVerification"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audit/pre-decisions": {
      "get": {
        "summary": "List pre-decisions, newest first",
        "tags": [
          "audit"
        ],
        "description": "For the lender's own compliance staff. Nobody outside the organisation can list or replay pre-decisions.",
        "parameters": [
          {
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreDecisionList"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audit/pre-decisions/{applicationId}/replay": {
      "post": {
        "summary": "Re-run the engine against the stored inputs",
        "tags": [
          "audit"
        ],
        "description": "Uses the STORED application, the STORED bureau lookup, the policy version RECORDED ON THE\nPRE-DECISION, and `submittedAt` as the clock — never today's policy, never a fresh bureau call,\nnever today's wall clock. The clock matters: `screen` derives age at maturity from it, so\nreplaying with now() would turn an applicant who was 74 at maturity into an age decline and\nreport tampering where there is none.\n\nIt compares the ENGINE's verdict, never the composed outcome. A human who overrode a referral\ndid not tamper with anything, and flagging them would make every legitimate override look like\nfraud. An `engineVersion` difference is reported but does not make `match` false on its own —\nthat is what the column is for.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "applicationId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Replay"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "https://loan-predecision.onrender.com",
      "description": "Deployed instance (free tier; first request after a quiet period is slow)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "tags": [
    {
      "name": "health",
      "description": "Liveness and readiness. They answer different questions."
    },
    {
      "name": "operations",
      "description": "Metrics. Behind the auditor scope, not public."
    },
    {
      "name": "applications",
      "description": "Submit an application and read its status."
    },
    {
      "name": "reviews",
      "description": "Record the outcome of a manual review."
    },
    {
      "name": "audit",
      "description": "Evidence: the event chain, and replay of a stored pre-decision."
    }
  ]
}